fakeHorse updates
fakeHorse follows the player now. originally i was going to instantiate a new horse behind the player after the horse was caught, but instead i decided to try and make it so we could just keep the already generated horse to save on optimization. it works for the most part, just need to refine it works with multiple horses (instead of just one horse). current issue: if any horse is captured, all of the horses follow player.